+++ src/data/search-urls.js
@@ -47,3 +47,11 @@ function returnShowOnMapUrl(query, shouldEncode = true) {
         case 'yandexmaps': return `https://yandex.ru/maps/geo/${encodedQuery}`; break;
     }
 }
+
+function returnTranslateUrl(query) {
+    switch (configs.preferredTranslateService) {
+        case 'google': return `https://translate.google.com/?sl=auto&tl=${configs.languageToTranslate}&text=${encodeURI(query.trim())}`; break;
+        case 'yandex': return `https://translate.yandex.ru/?lang=auto-${configs.languageToTranslate}&text=${encodeURI(query.trim())}`; break;
+        case 'bing': return `https://www.bing.com/translator?from=auto&to=${configs.languageToTranslate}&text=${encodeURI(query.trim())}`; break;
+    }
+}